+++ components/mdx/index.ts
@@ -1,0 +1,3 @@
+export { Callout } from './Callout';
+export { CodeBlock } from './CodeBlock';
+export { YouTubeEmbed } from './YouTubeEmbed';